requirements.txt
eth-brownie
redis